Program finder updates
