+++ includes/head.html
@@ -8,6 +8,7 @@
   <title>University of Missouri - Kansas City</title>
   <!-- <link rel="stylesheet" href="/assets/css/print.css" type="text/css" media="print" /> -->
   <meta content="#0066cc" name="theme-color">
+  <meta name="alert-block" content="84fe7bf3ac1e04cd75402c87af2d0f19" />
   <style>
     .hero{
       background-image: url('/src/images/kansas-city-skyline-downtown.webp');
