+++ includes/ks/directory-standard.html
@@ -8,63 +8,10 @@
         <label class="form__label">Search</label>
         <input name="s" class="form__text" id="dirSearch" type="search">
       </div>
-      <div class="form__group">
-        <label class="form__label">Categories</label>
-        <div class="form__sw">
-          <input name="c" id="dirSearchCat" class="form__select" list="categories">
-          <datalist id="categories">
-            <option></option>
-            <option value="Test">Test</option>
-            <option value="Thing">Thing</option>
-            <option value="Another">Another</option>
-          </datalist>
-        </div>
-      </div>
     </div>
   </form>
 </div>
 
-<div class="di">
-  <div class="di__item" data-filter=""> <!-- Grid layout -->
-    <img src="https://picsum.photos/90/90" class="di__image" alt="" />
-    <p class="di__name"><a href="#">First Name</a></p>
-    <p class="di__title">Faculty Member<br /></p>
-    <p class="di__dept">Social Sciences</p>
-    <p class="di__au">SHSS</p>
-  </div>
-  <div class="di__item" data-filter=""> <!-- Grid layout -->
-    <img src="https://picsum.photos/90/90" class="di__image" alt="" />
-    <p class="di__name"><a href="#">First</a></p>
-    <p class="di__title">Title</p>
-    <p class="di__dept">Depart the unversity</p>
-    <p class="di__au">computer science and even more departments!</p>
-  </div>
-  <div class="di__item" data-filter=""> <!-- Grid layout -->
-    <img src="https://picsum.photos/90/90" class="di__image" alt="" />
-    <p class="di__name"><a href="#">First Last;fdsal;khfdsl;kjfdsal;jfdsa</a></p>
-    <p class="di__title">Title<br />Title 2fdsafdsafdsafdsa</p>
-    <p class="di__dept">Department of the longest name at the unversity</p>
-    <p class="di__au">s!</p>
-  </div>
-  <div class="di__item" data-filter=""> <!-- Grid layout -->
-    <img src="https://picsum.photos/90/90" class="di__image" alt="" />
-    <p class="di__name"><a href="#">First</a></p>
-    <p class="di__title">Title</p>
-    <p class="di__dept">Depart the unversity</p>
-    <p class="di__au">computer science and even more departments!</p>
-  </div>
-  <div class="di__item" data-filter=""> <!-- Grid layout -->
-    <img src="https://picsum.photos/90/90" class="di__image" alt="" />
-    <p class="di__name"><a href="#">First Last;fdsal;khfdsl;kjfdsal;jfdsa</a></p>
-    <p class="di__title">Title<br />Title 2fdsafdsafdsafdsa</p>
-    <p class="di__dept">Department of the longest name at the unversity</p>
-    <p class="di__au">s!</p>
-  </div>
-  <div class="di__item" data-filter="test"> <!-- Grid layout -->
-    <img src="https://picsum.photos/90/90" class="di__image" alt="" />
-    <p class="di__name"><a href="#">First</a></p>
-    <p class="di__title">Title</p>
-    <p class="di__dept">Depart the unversity</p>
-    <p class="di__au">computer science and even more departments!</p>
-  </div>
+<div class="di paginationDirectory"></div>
+<div id="pagination">
 </div>
