+++ includes/ks/hero-hp.html
@@ -1,6 +1,9 @@
 <div class="hero hero__hp">
   <div class="hero__content">
-    <svg class="hero__college" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 797 361">
+    <div>
+      <span class="hero__title">R1 Proud!</span>
+    </div>
+    <!-- <svg class="hero__college" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 797 361">
       <defs>
         <clipPath id="clip-path">
           <rect width="797" height="361" transform="translate(-659 371)" fill="#06c"/>
@@ -38,7 +41,7 @@
           </g>
         </g>
       </g>
-    </svg>
+    </svg> -->
     <div class="hero__text">
       <p class="hero__subhead text--upper mask">
           <span>Like the future is yours</span>
