+++ includes/pf/ms_filter.html
@@ -1,5 +1,6 @@
 <div class="filter hide-filter-area" id="filter-toggle" aria-hidden="true">
   <h3 class="pf-results__title">Filter By:</h3>
+    <p>Filters automatically apply, close this window to see your updated results.</p>
     <p class="pf-results__heading">Program Type</p>
     <div id="facets--type"><div class="ais-RefinementList"><ul class="ais-RefinementList-list"><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Graduate</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Undergraduate</label></div></div></li></ul></div></div>
