+++ includes/pf/ms_search.html
@@ -9,12 +9,14 @@
           <span class="ais-SearchBox-loadingIndicator" hidden=""><svg aria-label="Results are loading" class="ais-SearchBox-loadingIcon" width="16" height="16" viewBox="0 0 38 38" stroke="#444" aria-hidden="true"><g fill="none" fill-rule="evenodd"><g transform="translate(1 1)" stroke-width="2"><circle stroke-opacity=".5" cx="18" cy="18" r="18"></circle><path d="M36 18c0-9.94-8.06-18-18-18"><animateTransform attributeName="transform" type="rotate" from="0 18 18" to="360 18 18" dur="1s" repeatCount="indefinite"></animateTransform></path></g></g></svg></span>
         </form>
         <div class="pf-search__view-all">
-          <button class="button button--umkc-navy button--outline" href="#" target="_blank">View All Programs</button>
+          <button class="o-btn button button--umkc-navy button--outline" data-toggle="#interests">
+            <span class="show">Show</span>
+            <span class="hide">Hide</span> Interest(s)
+          </button>
         </div>
       </div>
     </div>
-
-    <div class="pf-search__interests">
+    <div class="pf-search__interests" id="interests">
       <h3 class="pf-search__title">Choose Your Interests</h3>
       <div class="ais-RefinementList"><ul class="ais-RefinementList-list"><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Agriculture, Food and Natural Resources (7)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Architecture and Construction (7)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Arts, Audio/Video Technology and Communications (70)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Business, Management and Administration (29)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Education and Training (53)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Finance (7)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Government and Public Administration (10)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Health Science (42)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Hospitality and Tourism (3)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Human Services (21)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Information Technology (5)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Law, Public Safety, Corrections and Security (7)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Manufacturing (1)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Marketing, Sales and Service (5)</label></div></div></li><li class="ais-RefinementList-item"><div><div class="filter-input-wrapper"><label><input type="checkbox"> Science, Technology, Engineering and Mathematics (81)</label></div></div></li></ul></div>
     </div>
