+++ main.js
@@ -10,6 +10,7 @@ import '/src/js/tabs.js'
 import '/src/js/quotes.js'
 import '/src/js/video.js'
 import '/src/js/directory.js'
+import '/src/js/directory-pagination.js'
 import '/src/js/resource.js'
 import '/src/js/alerts.js'
 import '/src/js/replaceMe.js'
@@ -17,6 +18,7 @@ import '/src/js/infopopup.js'
 import '/src/js/video-shorts.js'
 import '/src/js/count-up.js'
 import '/src/js/consent.js'
+import '/src/js/pfinder-autocomplete.js'
 // import '/src/js/program-finder.js'
 // import '/src/js/pfinder-compare.js'
