+++ pfinder-search.html
@@ -8,7 +8,7 @@
 
 		{{> pf/header }}
 		<div id="landing-page" class="main-content">
-			{{> hero standard="true" }}
+			<!-- {{> hero standard="true" }} -->
       <!-- Meilisearch, form and interests list -->
       {{> pf/ms_search }}
