+++ src/js/alerts.js
@@ -74,7 +74,7 @@ async function showAlert(data){
     // If active and hpOnly is true and the location is the homepage, assigned for UMKC
 
     let matches = /http(s)?:\/\/(www\.)?umkc\.edu\/(index\.html)?/
-    if ( !sessionAlertHide && ( active || ( active && hpOnly && matches.test(window.location.href) ) ) ) {
+    if ( !sessionAlertHide && ( active && hpOnly && matches.test(window.location.href) || ( !hpOnly && active) ) ) {
       let content = current[2].text;
       alertDiv.innerHTML = alertDiv.innerHTML.replace("{{content}}", content);
       mainContainer.prepend(alertDiv);
