+++ src/scss/04-layouts/_program-finder.scss
@@ -35,6 +35,7 @@
 
     &__buttons{
       margin-top: 2rem;
+      text-align: center;
     }
     &__form-wrapper {
         max-width: 800px;
